added the final code.

+++ src/main/java/com/claim/claim_processing/integration/loanAdjustment/dto/LoanDetailResponseDto.java
@@ -1,6 +1,8 @@
 package com.claim.claim_processing.integration.loanAdjustment.dto;
 
 
+import java.math.BigDecimal;
+
 import lombok.*;
 
 @Getter
@@ -12,6 +14,6 @@ public class LoanDetailResponseDto {
 
     private String loanName;
     private String status;
-    private Double loanAmount;
-    private Double outstandingAmount;
+    private BigDecimal loanAmount;
+    private BigDecimal outstandingAmount;
 }
